refactor(openomni): split subagent runtime internals
